useLayoutEffect how it works
